GymBro/backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
pydantic
python-dotenv
#psycopg[binary]  # Not strictly needed for SQLite but good to have if we switch back
